ResumeExtractor/requirements.txt
fastapi
uvicorn
spacy
python-docx
PyPDF2
pytesseract
pillow
fuzzywuzzy
scikit-learn
python-Levenshtein
docx2txt